move users table to results
