+++ packages/front-end/components/Experiment/TabbedPage/ExperimentHeader.tsx
@@ -3,11 +3,11 @@ import {
   ExperimentPhaseStringDates,
 } from "back-end/types/experiment";
 import Link from "next/link";
-import { FaHome, FaUsers } from "react-icons/fa";
+import { FaHome } from "react-icons/fa";
 import { PiChartBarHorizontalFill } from "react-icons/pi";
 import { useRouter } from "next/router";
 import { getAffectedEnvsForExperiment } from "shared/util";
-import React, { useMemo, useState } from "react";
+import React, { useState } from "react";
 import { date, daysBetween } from "shared/dates";
 import { MdRocketLaunch } from "react-icons/md";
 import { SDKConnectionInterface } from "back-end/types/sdk-connection";
@@ -24,13 +24,10 @@ import TabButton from "@/components/Tabs/TabButton";
 import usePermissions from "@/hooks/usePermissions";
 import HeaderWithEdit from "@/components/Layout/HeaderWithEdit";
 import Modal from "@/components/Modal";
-import Tooltip from "@/components/Tooltip/Tooltip";
 import { useScrollPosition } from "@/hooks/useScrollPosition";
 import ResultsIndicator from "../ResultsIndicator";
-import { useSnapshot } from "../SnapshotProvider";
 import { StartExperimentBanner } from "../StartExperimentBanner";
 import ExperimentStatusIndicator from "./ExperimentStatusIndicator";
-import OverflowText from "./OverflowText";
 import StopExperimentButton from "./StopExperimentButton";
 import { ExperimentTab, LinkedFeature } from ".";
 
@@ -54,11 +51,6 @@ export interface Props {
   editTargeting?: (() => void) | null;
   editPhases?: (() => void) | null;
 }
-
-const shortNumberFormatter = Intl.NumberFormat("en-US", {
-  notation: "compact",
-  maximumFractionDigits: 1,
-});
 
 export default function ExperimentHeader({
   tab,
@@ -103,8 +95,6 @@ export default function ExperimentHeader({
 
   const [startExperiment, setStartExperiment] = useState(false);
 
-  const { analysis } = useSnapshot();
-
   const canCreateAnalyses = permissions.check(
     "createAnalyses",
     experiment.project
@@ -119,19 +109,6 @@ export default function ExperimentHeader({
     }
   }
   const canRunExperiment = canEditExperiment && hasRunExperimentsPermission;
-
-  const [totalUsers, variationUsers] = useMemo(() => {
-    let totalUsers = 0;
-    const variationUsers: number[] = [];
-    analysis?.results?.forEach((dim) => {
-      dim?.variations?.forEach((v, i) => {
-        totalUsers += v.users;
-        variationUsers[i] = variationUsers[i] || 0;
-        variationUsers[i] += v.users;
-      });
-    });
-    return [totalUsers, variationUsers];
-  }, [analysis]);
 
   return (
     <>
@@ -407,66 +384,8 @@ export default function ExperimentHeader({
                 />
               </TabButtons>
             </div>
-            <div className="flex-1" />
 
-            {experiment.status !== "draft" && totalUsers > 0 && (
-              <div className="col-auto mr-2 users">
-                <Tooltip
-                  usePortal={true}
-                  body={
-                    <table className="table my-0">
-                      <thead>
-                        <tr>
-                          <th className="border-top-0">Variation</th>
-                          <th className="border-top-0">Users</th>
-                        </tr>
-                      </thead>
-                      <tbody>
-                        {experiment.variations.map((v, i) => (
-                          <tr key={i}>
-                            <td
-                              className={`variation with-variation-label variation${i}`}
-                            >
-                              <div className="d-flex align-items-center">
-                                <span
-                                  className="label"
-                                  style={{
-                                    width: 20,
-                                    height: 20,
-                                  }}
-                                >
-                                  {i}
-                                </span>{" "}
-                                <OverflowText
-                                  className="font-weight-bold"
-                                  maxWidth={150}
-                                  title={v.name}
-                                >
-                                  {v.name}
-                                </OverflowText>
-                              </div>
-                            </td>
-                            <td>
-                              {shortNumberFormatter.format(
-                                variationUsers[i] || 0
-                              )}
-                            </td>
-                          </tr>
-                        ))}
-                      </tbody>
-                    </table>
-                  }
-                >
-                  <div className="px-2 py-1 rounded text-gray">
-                    <FaUsers />{" "}
-                    <code className="text-dark">
-                      {shortNumberFormatter.format(totalUsers)}
-                    </code>{" "}
-                    users
-                  </div>
-                </Tooltip>
-              </div>
-            )}
+            <div className="flex-1" />
             <div className="col-auto experiment-date-range">
               {startDate && (
                 <>
